# pipeline/api-specs/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    IngestOperationRestoreIn:
      properties:
        cancel_operation_ids:
          items:
            type: string
          title: Cancel Operation Ids
          type: array
        deleted_pages:
          items:
            type: string
          title: Deleted Pages
          type: array
        operation_id:
          title: Operation Id
          type: string
        rebuild_pages:
          items:
            $ref: '#/components/schemas/RebuildPageIn'
          title: Rebuild Pages
          type: array
        restore_to_operation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Restore To Operation Id
        source_page:
          $ref: '#/components/schemas/SourceSnapshotRestoreIn'
        workspace_id:
          title: Workspace Id
          type: string
      required:
      - operation_id
      - workspace_id
      - rebuild_pages
      - restore_to_operation_id
      - cancel_operation_ids
      - source_page
      title: IngestOperationRestoreIn
      type: object
...
    PipelineRunOut:
      properties:
        log_path:
          title: Log Path
          type: string
        manifest:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Manifest
        output_dir:
          title: Output Dir
          type: string
        run_id:
          title: Run Id
          type: string
        status:
          title: Status
          type: string
      required:
      - run_id
      - status
      - output_dir
      - log_path
      title: PipelineRunOut
      type: object
...
info:
  title: Fruition Pipeline Lab API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /chat-wiki/runs:
    post:
      operationId: run_chat_wiki_endpoint_chat_wiki_runs_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatWikiRunIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineRunOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Run Chat Wiki Endpoint
      tags:
      - wiki-ingest
...
  /pipeline/reingest-runs:
    post:
      operationId: run_reingest_pipeline_endpoint_pipeline_reingest_runs_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReingestRunIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineRunOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Run Reingest Pipeline Endpoint
      tags:
      - wiki-ingest
  /pipeline/runs:
    post:
      operationId: run_pipeline_endpoint_pipeline_runs_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineRunIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineRunOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Run Pipeline Endpoint
      tags:
      - wiki-ingest
  /pipeline/runs/{run_id}:
    get:
      operationId: get_pipeline_run_pipeline_runs__run_id__get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Pipeline Run Pipeline Runs  Run Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Pipeline Run
      tags:
      - wiki-ingest
  /pipeline/runs/{run_id}/logs:
    get:
      operationId: get_pipeline_logs_pipeline_runs__run_id__logs_get
      parameters:
      - in: path
        name: run_id
        required: true
        schema:
          title: Run Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            text/plain:
              schema:
                type: string
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Pipeline Logs
      tags:
      - wiki-ingest
...
  /wiki/documents/{document_id}/context:
    get:
      operationId: get_document_wiki_context_wiki_documents__document_id__context_get
      parameters:
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Document Wiki Context Wiki Documents  Document Id  Context Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Document Wiki Context
      tags:
      - wiki
  /wiki/graph:
    get:
      operationId: get_wiki_graph_wiki_graph_get
      parameters:
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Wiki Graph Wiki Graph Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Wiki Graph
      tags:
      - wiki
  /wiki/ingest-restore-runs:
    post:
      operationId: restore_ingest_operation_wiki_ingest_restore_runs_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestOperationRestoreIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Restore Ingest Operation Wiki Ingest Restore Runs Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Restore Ingest Operation
      tags:
      - wiki
  /wiki/lint-restore-runs:
    post:
      operationId: restore_lint_operation_wiki_lint_restore_runs_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LintOperationRestoreIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Restore Lint Operation Wiki Lint Restore Runs Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Restore Lint Operation
      tags:
      - wiki
  /wiki/maintenance/lint:
    post:
      operationId: lint_wiki_workspace_wiki_maintenance_lint_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiLintIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiLintOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Lint Wiki Workspace
      tags:
      - wiki
  /wiki/pages/lookup:
    post:
      operationId: lookup_wiki_pages_wiki_pages_lookup_post
      parameters:
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiPageLookupIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response Lookup Wiki Pages Wiki Pages Lookup Post
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Lookup Wiki Pages
      tags:
      - wiki
  /wiki/pages/{page_id}:
    get:
      operationId: get_wiki_page_wiki_pages__page_id__get
      parameters:
      - in: path
        name: page_id
        required: true
        schema:
          title: Page Id
          type: string
      - in: query
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Wiki Page Wiki Pages  Page Id  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Wiki Page
      tags:
      - wiki
  /wiki/pages/{page_id}/rename:
    patch:
      operationId: rename_wiki_page_wiki_pages__page_id__rename_patch
      parameters:
      - in: path
        name: page_id
        required: true
        schema:
          title: Page Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WikiPageRenameIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Rename Wiki Page Wiki Pages  Page Id  Rename Patch
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Rename Wiki Page
      tags:
      - wiki
  /wiki/workspaces/{workspace_id}/documents/{document_id}:
    delete:
      operationId: delete_document_wiki_data_wiki_workspaces__workspace_id__documents__document_id__delete
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: path
        name: document_id
        required: true
        schema:
          title: Document Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Document Wiki Data
      tags:
      - wiki
  /wiki/workspaces/{workspace_id}/last-updated:
    get:
      operationId: get_last_wiki_updated_wiki_workspaces__workspace_id__last_updated_get
      parameters:
      - in: path
        name: workspace_id
        required: true
        schema:
          title: Workspace Id
          type: string
      - in: header
        name: X-Internal-Token
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Internal-Token
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Last Wiki Updated Wiki Workspaces  Workspace Id  Last Updated Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Last Wiki Updated
      tags:
      - wiki